First correct compilation of programm
